+++ core/helpers.py
@@ -133,8 +133,6 @@ class CompanyParser(great_components.helpers.CompanyParser):
         data.setdefault('expertise_products_services', {})
         data.setdefault('expertise_countries', [])
         data.setdefault('expertise_industries', [])
-        if settings.FEATURE_FLAG_HARD_CODE_USER_INDUSTRIES_EXPERTISE:
-            data['expertise_industries'] = ['FOOD_AND_DRINK']  # food and drink
         super().__init__(data=data)
 
     def __getattr__(self, name):