+++ exportplan/core/helpers.py
@@ -1,10 +1,14 @@
 import pytz
+from django.conf import settings
+from hashids import Hashids
 from iso3166 import countries_by_alpha3
 
 from core import models
 from core.templatetags.content_tags import format_timedelta
 from directory_api_client import api_client
 from exportplan.core.processor import ExportPlanProcessor
+
+hashids = Hashids(settings.HASHIDS_SALT, min_length=8)
 
 
 def create_export_plan(sso_session_id, data):
@@ -127,6 +131,7 @@ def get_exportplan_detail_list(sso_session_id):
         # On list page we need to know sections complete only EP processor can calculate this
         # Move this to an easy method TODO
         ep['calculated_progress'] = ExportPlanProcessor(ep).calculate_ep_progress()
+        ep['url'] = ExportPlanProcessor(ep).get_absolute_url
 
     return exportplan_list
 
@@ -135,3 +140,23 @@ def get_exportplan(sso_session_id, id):
     response = api_client.exportplan.detail(sso_session_id=sso_session_id, id=id)
     response.raise_for_status()
     return response.json()
+
+
+def h_encode(id):
+    return hashids.encode(id)
+
+
+def h_decode(h):
+    z = hashids.decode(h)
+    if z:
+        return z[0]
+
+
+class HashIdConverter:
+    regex = '[a-zA-Z0-9]{8,}'
+
+    def to_python(self, value):
+        return h_decode(value)
+
+    def to_url(self, value):
+        return h_encode(value)